ISSUE-026 Server: Implemented "friendly" rounding. This provides an additional alternative to the normal rounding (that is concerned with decimal places) by rounding values so that only 3 digits are shown that aren't zero.
It is flexible but currently it is set to 3 digits. As in, if a conversion result rounded is 501.25, it'll be 501.25. But if the value is 1000 or greater, it'll start rounding to whole numbers in the 1's place because it cannot be expected for a student to worry about more digits than that.

Ex: 1.1001 km => 1100.1 m (exact/rounded) => 1100 m (friendly)
Ex: 11.001 km => 11,001 m (exact/rounded) => 11,000 m (friendly)