ISSUE-044 Server: Wrote setActiveCourse() that sets all student courses to Inactive and the targeted course to active.
It is slightly smart and won't attempt to do updates that won't change anything.

Added a util function setStatusForCourses() with the first case of the new error GraphQlDumpWarning. This is necessary because running a query or mutation where the OR statement is empty will run on everything, at the very least returning everything in a query and at the worst changing everything in a mutation. 😱

New error StudentNotEnrolled because that makes sense.