ISSUE-020 Server: Fixed an issue where Questions with answered Surveys were being processed twice.
This is due to the fact that the if statement this part was in originally had return statements and because my linter settings insist I don't use unnecessary else statements when a return is in it this happened! Haha, code cleanliness leads to me making a mistake!

Moved around comments to be above the for loop that's being described.

Being less stupid and using Array.splice() instead of Array.slice().