ISSUE-038 Server: Added Schema support for Survey status and logic support for it as well.
Added tests for Status in Surveys and also updated the README.

If a user has skipped the Survey question a new status field is now available to be viewed. If there is a no Survey it will be null like response.

I wish to note that Survey questions that are skipped WILL take up space in the final list. The plan is that these skipped questions will be ignored by the client. So beware around very small challenges (like... one), as that one question could be a skipped Survey. Left a TODO note about it in challengeGenerator.js.

Making skipped Surveys not have any detail if present to make things a little more consistent.