ISSUE-039 Server: Wrote addChallengeResults mutation.
Hoooo boy! I'm really proud of this work! I made something clean and efficient and flexible and it works without much drama! Now I've completed the two sides: The query getChallenge and the mutation addChallengeResults.

I'm a meticulous comment writer so I don't think I need to explain much here, just take a look at the new function I wrote.

I noticed I didn't need Survey statuses in any of my Course queries so I removed that.

Added an if-statement in surveysScoreUpdateClauseGenerator() that would handle the situation where the addChallengeResults() would have Surveys belonging to the Course that we didn't intend to adjust the scores for.