ISSUE-040 Server: Defined new Question row "author".
Defined new User row "questionsWritten" and "questionsReviewed".

New Questions statuses defined in the constants for when they're in review or rejected. (Approved is that the Question is set to Active).

Little typo fix in feedback.js.

Changed name of User row feedbackGiven to feedbackWritten to match the other rows I just made.