Server: Single character mistake fixed.
I was writing question syntax for step as if they were accuracy.
For example: "This is the height of a man. [1.6,1.9m(0.05)a]"
Should be: "This is the height of a man. [1.6,1.9m(0.05)s]"

It wasn't breaking anything as the system will default to a step of 1.0 if it the syntax wasn't found.