ISSUE-082 Client: Switching UserDetails Component to use generic user() query.
This simplifies logic so I don't have to switch Queries depending if the page being viewed is /user/me or /user/<userid>.

Using fragments like crazy to get a bunch more data.

Server: Adjusted query user() to allow anyone to use it as an alternative to me() so long as students and teachers are only looking at themselves (I'll need to make it so teachers can look at their students later).

Updated README to explain this change.