ISSUE-114 Client: Using the new-and-improved withAuth HOC to provide userTokenData to UserDetailEnrollment and UserDetailCourse.
I use it to prevent users besides moderators and admins from being able to see the buttons to enroll and assign a course for a student. This wouldn't happen right now, but it's a future-proof against, I dunno, when I write the ability for teachers to view their students' pages.

With this feature all I need to do is add a withAuth wrapper and, boom, instant access to userTokenData which will allow me to grab all sorts of useful things.

Also added a new fast-start script for the time you know what you're doing any want to skip the SemanticUI CSS build process.