ISSUE-113 Client: Implemented an alternative behavior for Login/Signup using withAuth.
If the user is detected as logged in they will see a different page, AlreadyLoggedIn.js, that'll tell them to log out or go to the homepage.

Switched Login.js to use PureComponent because it uses just one state variable.

Added the ability for Logout to redirect back to where it came from if the state.from.pathname is defined.