ISSUE-105 Client: Initial building of UserDetailBasicsEditor.js
It currently isn't working due to unacceptable null values in the mutation (I imagine I'll be able to fix them.

Buttons to turn on and off edit mode are hacky and need to be replaced as having them inside QueryHandler causes warnings because QueryHandler passes prop queryData to <button> elements.

Defined a really sweet function for utils called userDetailFormValidator() that uses a lot of smart tricks to provide a way to validate user data forms (login, signup, and this new edit page).

Installed lodash. Needed it for mergeWith(). The way I use it in userDetailFormValidator() is actually a bit complicated as I needed it to support recursively merging objects together. It totes works, though!!

Switched userid prop to userId for User/UserDetails - all lowercase is only for GraphQL arguments. Gotta maintain camelCase for everything else.

Already wrote the update function for the edit mutation, though I haven't actually tested it yet!