ISSUE-105 Client: Adjustment to QueryHandler and restructure of UserDetailBasicsEditor.
Moved the contents of UserDetailBasicsEditor to UserDetailBasicsEditorForm to match the set-up to be much more like the UserDetailCourseAssign and UserDetailEnroll where the query is in the parent, and the mutation is by itself in sub component, and finally at the bottom is the form with the validation and the simple call to submit the info.

UserDetailBasicsEditor now is in control of the Apollo Mutation component, greatly cleaning up clutter from UserDetails component.

Adjusted QueryHandler to also pass on the prop 'query' to all its children, because mutations rely on this query object (a gql object, thus the eslint exception) to make updates when complete, referencing the constant in the same component where the mutation lives was repetitive, messy, and prone to mistakes.