ISSUE-108 Client: Initial Subject listing with Query.
New queries, one for the case where it's a public display of Subjects, and one where it's for a student (it'll have mastery information on display).

Renamed User.js to UserPage.js to be a little more obvious the meaning behind it.

Made the openEditor button for UserDetailBasics optional. For example, if a Teacher is looking at a student they will probably not have the option to edit their stuff!