ISSUE-108 Client: Switching from publicList prop to studentId.
This communicates the same data we needed and is also more useful than a simple boolean.

Added yet another Component, Mastery.js. The idea is that I might make it pull double duty: Both to display as a child of a SubSubject, and also can have a SubSubject component as child - and through props will determine whether it's one or the other.

That's a cool thing about GraphQL and how you can be cyclical with relations, you can have one or the other:
SubSubject.masteries[0].id
Mastery.subSubject.id