ISSUE-123 Client: Wrote navigateObjectDots() and adjusted cache update functions.
The navigateObjectDots() function basically allows me to search through a complex object using dot notation in a string (or array, as required by the recursive nature) in order to do more complicated cache updates. The case was using cachePushIntoArray() in UserDetailCourseAssign where I knew the studentId and had to push into enrollment.courses without knowing the enrollment ID. So, I wondered how to gain access to that array programatically with an address like "enrollment.courses" - that's when I wrote this function.

I went back and updated cacheNewObject() and cachePushIntoArray() to use navigateObjectDots(). But don't be surprised if I add this kind of address support to cacheUpdateObject(), cacheDeleteObject() at some point - for example, wanting to delete a student's enrollment object while only knowing the studentId and not the enrollmentId. But for now, I think I could get away by using findRecursive() and doing something like `delete target.enrollment` in that case.

Updated remaining mutations to use the cache update functions.

Ran into a big issue with the mergeWith() and the mergeCustomizer() functions where for some reason if the obj and the src were both null, the resulting set value would be an empty object. In attempting to figure it out I brought in the lodash function isPlainObject() to check and that ended up solving things. Now, that should mean that arrays will no longer merge, but for the time being I don't think that's much of a concern since I normally treat arrays as entire objects that need to be replaced, since there isn't much meaning behind the indexes of values in an array, I doubt I'll be replacing "value at index [3]" or anything like that.