ISSUE-124 Client: SubSubject component simplified and no longer attempts to display Mastery data inside it.
The parent components of SubSubject + Mastery (SubSubjectsList) and Mastery + SubSubject (UserDetailMasteries) are now in charge of deciding if the data is shown.

Depending on whether the required props are passed in, they'll show or not show Assign and Enable/Disable buttons, making things a little more flexible and easier to use.

Removed repeated studentId prop from SubjectsList, Subject, SubSubjectsList, and SubSubject. The studentId was available in the variables object passed through with queryInfo and is thus passed into MasteryAssign from SubSubjectsList.

Switched the public query to use allSubjects query. Smarter handling around the Public vs Student queries in Subjects.

UserDetails now has the fetch policy to always grab by network. It still writes to cache so that I can make cache updates on mutations, but it'll always hit the server to guarantee that the page always displays perfectly fresh data.

Server: Made query allSubjects public so that the Subjects and SubSubjects can be viewed by users who are not logged in.

Updated README for change.