ISSUE-107 Client: Masteries can be displayed sorted by Subject.
This is some weird stuff! I flip around the object tree from Mastery>SubSubject>Subject to Subject>SubSubject>Mastery depending on a prop set in UserDetailMasteries: organizeBySubject and subjectCompactView.

New function reverseMasteriesData() in UserDetailMasteries does the weird work with the help of two new utils functions cacheTargetExists() and rootCopy().

Right now the Mastery details in the SubSubject accordion title is blank, and I want to be able to sort Subjects and SubSubjects so that the order is always the same for everyone.