ISSUE-104 Client: Improvements to SurveySerious reengineering of QueryHandler.
I realized that if I went to a user page with a non-existant user ID that I'd get page errors. So I needed to design a way QueryHandler to recognize when a query resulted in nothing. This required defining a new util function isEmptyRecursive because lodash's isEmpty would be tricked by the queryData `data.user === null`. I wanted a function that would look at that kind of object and say "no, that IS EMPTY!" because lodash's isEmpty simply looked at whether or not an object had keys.

Added a link to the /subjects page below masteries in CourseAndMasteriesAndSurveys component.

SurveysList now displays a truncated version of the survey's question. This required writing two new utils functions questionTextGrabber() and stringTruncator().

Change to the Welcome page to direct to another user (since I had to reset the other day).

Converted the old FloatingCenterGrid component to a functional component.

Lower-cased user type and status names, now will use utils.firstLetterCap() when necessary.

Switched UserDetails component to use a constructor() function to get rid of annoying syntax errors.