ISSUE-146 Client: Fixed MasteryAssign issue for Subjects page.
It involved first fixing a bug in isEmptyRecursive (I was accidentally marking non-empty objects as empty, see previous commit). This but would break the Subjects page if a student had no Masteries set.

From there I made it so that the allSubjects query added an empty and formatted array for Mastery information in the masteries field. Thankfully first: 0 worked exactly as I needed it.

Discovered some nuance about Apollo's cache, where if you update the cache with more information than the original query requested (for example, when figuring out the bug my query alLSubjects only asked for the Mastery's ID) the cache will drop the unwanted information (in my case, status and score).

Because I needed the same data in multiple places, I defined a new fragment "MasteryDataAllExtra" and used it in multiple places.