ISSUE-141 Client: First rough draft of the QaReviewBasics component.
Defined a new file, propTypes.js, that'll hold the gigantic PropType shape of the Qa Object.

Added the deline package, this allows me to use template strings and add linebreaks to them so they aren't crazy long on a single line in my code but in the end will render as strings without linebreaks or extra spaces.

Added test links to the Welcome page to get me a written, a conversion, and a survey question ASAP.