ISSUE-141 Client: Fixed Apollo issue.
It was because I had the Survey's ID under the field name "id". This confused Apollo's heuristic matching, so I switched the name to "surveyId".

Also fixed some issues with the cache modification functions in utils where I wasn't taking into consideration the odds that the objects being searched through could be null.

Server: Updates to schema for the type QaSurveyResponseObject.