ISSUE-149 Server: Recovered mutation addSurveyAnswer() and updated it to use studentid instead of courseid.
Updated it so that you can set the score to something when answering a Survey. This is necessary to allow people to change their note (Survey.detail) without resetting their progress.

Mild improvement to surveysAnswerUpdatePayloadGenerator() in mutation/course.js. It'll now check to make sure that the input is correct later to save a loop.