ISSUE-149 Client: Have the rough draft of the Survey editor in place.
It's not nearly complete:
Does not do form validation.
Does not limit the min or max or step requirements I want.
It does not heed Question flags:
* `0x01` - "User detail note requested - optional"
* `0x02` - "User detail note requested - required"
Does not update the cache (will have to look into updating the cache modifier functions to look for "surveyId" instead of just "id").
Probably could be cleaned up in some places, we'll have to see.

Switched to addSurveyAnswer() mutation so I could get the Survey data back without having to do something insane like search for it inside a Course.