ISSUE-149 Client: Added "key" argument to cache modifier functions.
With the new "key" argument, you can now customize the object property key that gets checked. It can be as simple as the id having a different key name, or it can be a totally different value to be checked. Oh so exploitable.

Added another test to Test.js so show it working.

Erased unused import in SurveyAndQuestion component. WebStorm probably added it when I copy-pasted something around.