ISSUE-149 Client: QaReviewSurveyBasics now changes output depending on Question flags for detail requirements.
It will now say "None" in italics if there is an option for them but there is none.

QaReviewSurveyEditorForm also changes depending on the Question flags for detail requirements. It'll hide the note input if none is asked for (but will still show it if there is a note... and will disappear if you erase it, lol - I'm leaving that in).

Mild improvement to the utils function surveyAnswerValidator() to use more efficient helper function choiceWorder().

The issue from the previous commit is still there, though. Will change that tomorrow - don't wanna get started on it if it turns out to be a pain in the ass.