ISSUE-138 Server: Defined new mutation updateQuestion().
This was a bit of a brain teaser and involved a lot of code to get rolling.

Defined new utils function customMerge, like the one found in Client (with the new change made in the last commit).

Updates to README to the annotated QaObject section to be easier to read. I don't know what I was thinking by adding tons of white space thinking it'd be easier to comprehend...