ISSUE-138 Client: First basics of the QuestionReviewDetails component.
New limited data QA queries. Only includes data that's necessary to fill out the Question interface. This is the same data I use in the server-side Mutation updateQuestion() if you wonder where I got the inspiration for this.

Simplification of QaReview. No longer going to try and get a question editor here.