ISSUE-138 Client: Added Media row to QuestionReviewDetailsBasics.
I had no where to fit the image if defined so I added it to a Modal you can click on. Works great!

Example image of a dog I took a few years back.

Example links to the questions with the dog media.

Fix to a bad mistake in QaReviewBasics for the image.

Totally random change to QaReviewChoices. I didn't like how the horizontal lists behaved on line breaks.