ISSUE-138 Client: Aggressive streamlining of the whole question component family.
Each subsection now demands the props all organized in a row. No longer are they dealing with gross QA object data.

QuestionViewer is now the parent of all data as it is the first receiver of qaData from a query.

QuestionReview renamed to QuestionViewerLayout, which I believe makes more sense because it contains the grid layout.