ISSUE-138 Client: Tons of work on QuestionDetailsSubSubjectsSelector component. WIP but feature rich.
Decided to take a slick method of selecting SubSubjects. Instead of giving a list of every SubSubject belonging to a Subject in a row with them doubled up for every one that is toMetric and not I decided to do some client-side data manipulation and create a three-stage dropdown system for Subject>Scale>toMetric.

I'm thinking of somehow embedding dropdowns into SubSubjectReview as my next step.