ISSUE-138 Client: Initial, basic restructuring of SubSubjectReview component to rely less on a shoddy subSubjectData prop and instead on individual props.
Doing this required an additional field, parent.name, to the Subjects query I'm using, so that component QuestionDetailsSubSubject could call SubSubjectReview easier.

Strengthened utils function firstLetterCap() to be more flexible.

Disabled the lint rule "no-nested-ternary" because it's too useful in JSX to not use. It's hideously out of place in normal JS, but in JSX, where you can't use if, if/else, or switch cases, it's very useful.