ISSUE-138 Client: Writing some new cheater utils functions.
nod(), t0(), and t0t().

Also updated navigateObjectDots() after looking at it for a few minutes - I realized that if a null value was passed in it would throw an exception. The update is also slightly more sensible in construction by not having a second if statement and have a far clearer base case.