ISSUE-138 Client: Because some really obscure Apollo Client error would happen when the editor was closed immediately after a successful mutation in the case that the subSubject was modified I had to remove the use of LoadingButton's afterModalSuccess prop.
Updated the names of some props for LoadingButton pertaining to Modal configurations. Also, changed the order that the button props are applied to the trigger button for a Modal component overwrote buttonProps.disabled.

Added an "unsavedChanges" state to QuestionViewer component to track when we can close without needing a confirmation.

There is a wrinkle - when going into edit mode, likely due to QuestionDetailsSubSubjectsSelector, handleChange() is immediately triggered and marks the editor as having unsaved changes. One day I could stand to fix this but for now it still works better than not having it.