ISSUE-137 Client: Added unit copy button to QuestionDetailsAnswer component.
Can save a bunch of effort when writing multiple choice answers.

Also changed it so that setting any choice to written from a unit will make it blank immediately - this is to make the behavior consistent and to make sure that a user doesn't accidentally submit a numerical answer as a written one (they could do it intentionally, of course...).