ISSUE-172 Client: Added UserDetail modal to the table.
Unfortunately it doesn't update the details on the table after being complete (that'd take a mutation update function that could handle multiple query updates. While totally possible and maybe wildly cool, would be a huuuuge thing to do.)

Changed the import name of UserPage to match what it's actually called in App.js.