ISSUE-106 Client: Introduction of the MenuFrame component, a component that wraps around all components where a user is logged in and active.
Right now it's just a header menu that stretches across the top of the page. But eventually I want to see if I can make it as simple as a single button that will reveal a Sidebar when a button is clicked to save precious vertical screen space on mobile screens.

Reorganization of the whole routing table.

Also little change to UserListTable's modal (added a direct link to the user's /user/... page) and added all the colored buttons to Test.js.