ISSUE-177 Client: Added support for typing in negative numbers by allowing the first character to be "-".
If the user submits it then it'll fail the server-side entry by being an unacceptable value (non-float).

Fixed little bug in LoadingButton where the button text would flip to say "ERROR!" if it had an error in modal mode even when the prop buttonDisableOnError was set to false.

I did not add negative checking to survey entries as I will instead rely on the survey range instead.

Got rid of the backup Input components for Units in QuestionDetailsAnswer and QuestionDetailsQuestion and replaced them with messages saying to select a SubSubject first. It was confusing UI to say the least so I decided to remove it. I also removed the button for "Make First Choices" until the SubSubject was selected.

Server: Added rejection of negative values for incompatible units. Also updated some error messaging.

Defined a list of negative value compatible units (Fahrenheit and Celsius, as I plan to support temperature differences later on I decided to define a clean array of units instead of relying on a simple if statement).