ISSUE-187 Server: Updated qaGenerator.qaGenerate() to use a serialNumber to reliably provide a unique id property for each QA object.
This was necessary because generateChallenge would by design repeat conversion questions, resulting in multiple QA objects with the same id property when it was just "QA_<<questionid>>". Now it's "QA_###_<<questionid>>". ### is a 0-padded serial number starting at 0. I decided to leave in the questionid to prevent situations where the client could get confused between a bunch of QA_000's.

Updated unit tests.