ISSUE-191 Client: currentQa now has two pieces of data: answerData (the current user's input) and choices (the very common multiple choice selection).
All questions except for unanswered Surveys will have, possibly, a set of multiple-choice answers. Whether a conversion is raw input, a range, or multiple-choice, the options for multiple-choice are generated anyway with utils.choiceSelector().

Lots of wiggling around setting the data around currentQa. Had to use utils.t0() to make the showClearButton property work.

Switched ChallengeAnswerMultipleChoice to a functional component. Removed its state and constructor stuff and instead rely on ChallengeManager to determine the selected answers (given through the prop "choicesSelected") so that it can be saved to state and reloaded as needed.

New mascot temp image. Now he has sparkly eyes!