ISSUE-197 Server: Fixed issue with broken questionUpdate mutation.
Turns out in ISSUE-138 and commit 6e2bd8aa I had reduced some redundancy in the question-side of the Question submission schema (QuestionQuestionInput) by collapsing two range objects "conversioninput: ConversionQuestionInput" and "surveyrangeinput: RangeQuestionInput" (completely guilty of inconsistent naming for both the field names and the input names especially considering the two input objects were identical in structure!) to a new field "rangeinput" responsible for "[64,82f(1)s]".

Well, by chance there was another field named "conversioninput" in the sibling input (QuestionAnswerInput). In the code I went in and mistakingly renamed that field also to "rangeinput", which was totally wrong, as down the line in the functions checkAndParseQuestionAnswerInputs() and answerSyntaxFormatter() it was counting on there to be a property named "conversioninput" which was responsible for "[c(1)a]". It was now named "rangeinput". Bad!

So, I returned the name for the answerInput.rangeinput back to answerInput.conversioninput.

Additionally, I had renamed the questionInput.conversioninput to questionInput.rangeinput but had NOT set up a way to send questionInput.surveyrangeinput to checkAndParseQuestionAnswerInputs(). Which would pass off questionInput to questionSyntaxFormatter with questionInput.rangeinput.

By using nested ternary statements I've made questionInput.rangeinput have the conversion and survey ranges correctly collected. This bug would make it so that it would require me to re-define the rangeinput when making any modification to a survey question (I discovered it when trying to simply update a flag). Bad!