ISSUE-191 Client: 🎉 With this I hit super-ultra-official alpha state! 🎉
ChallengeResults submission works! So essentially this website has become functional. Very incomplete, but functional!

Wooo! Mark the date and time now. This is essentially the first rough draft of my website. It's not done but essentially the core functionality of the website is working!

Created new challenge state "challengeResultsSubmitted". It is set to true when the challenge's results are successfully received. This is used to notify the user, update the UI, and most importantly, prevent a user from re-submitting their results repeatedly to clone their mastery score progress.

I fixed an embarrassing divide-by-zero error in ChallengeList. I also used the same dilution tactic in the survey correct answer.