ISSUE-200 Server: Removed requirement for an answer on a Survey.
This had to be applied to both the datamodel (the actual Survey table) and the schema (where the answer is suggested to be required in QA objects).

The server logic of qaGenerator.js is sound. In the case that the answer is null the answer data becomes null without errors or weird behavior. So no additional changes are needed to patch holes in the logic.