ISSUE-199 Client: Cleaning up the style of ChallengeComplete component.
Added a "Rematch!" button at the bottom to encourage taking more Challenges.

Dealing with presentation issues where the label for the Progress components are getting a little squished. I changed the truncation limit to be 5 characters shorter and moved the value to the Progress component and limited the content of the label to have just the score change.

Server: Updated wording on more set questions. Starting with "Personally," is redundant as it is implied and a waste of space.