ISSUE-185 Client: Made menu items (except for logout) middle-clickable.
I was considering switching the Login/Signup buttons to be middle-clickable but they would lose their magic on redirecting backward to the Subjects page and I didn't want to lose that - besides, someone wouldn't be expected to middle-click it anyway...

Outside of scope... making Login component slightly prettier. Defined a new width setting "large" and use it for the Login component. Added a nice Icon to the Header component.