ISSUE-210 Client: Added incorrectAnswers tracking to replace the incorrectAnswerCount.
Yeaaaaah, this whole challenge system isn't all that clean. This is what happens when I don't bother to take the time to do design with diagramming for the most complex part of my website. But it appears to work just fine.

Fixed a weird auto-refactor problem in ChallengeMain.

Fixed a typo in ChallengeManager, the progress object should be skipped, not skip.