ISSUE-210 Client: Moved many parts of the currentChallenge state to challengeProgress state array.
This allows me to have settings in the client where it can be configured that choices and ranges be redrawn between repeats of the same question.

This involved a lot of ugly changes. Sorry for the diff gore.

Updated ChallengeCompleteDetailScores so that it shows a grey colored progress bar if the progress was negative. If positive it'll be in indicating mode where the color is handled.