ISSUE-223 Client: Created DocumentationPage component.
Converted my documentation to Semantic-UI-React'd HTML code. A lot of copy and pasting.

I split the content into JSX objects in a giant nested object. This way I can re-use existing material for other parts of the site.

Built new ExternalLink component that provides a secure and nice looking new-tab hyperlink to wherever.

Added link to docs to the menu. Still have room for it.