ISSUE-225 Client: Created DocumentationSubjectModal.
I am super happy with the results of this thing. I just wish I could've used an Accordion but it was bugged (at least the shorthand version, and I didn't want to try adding state to the whole thing). Not a huge loss as the whole thing is still eminently usable without them.

In DocumentationContent moved unit description preamble thing back to header as a subheader which works perfectly.