ISSUE-203 Client: Mascot responses in place and react to the streak.
Had to do a little bit of count cheating because the streak was what was going in to the question, not what it would be now.

Changed response details to the SubHeader.

Alignment is currently broken. Will need to figure out how to fix it...

Bonus:
Swapped the string "To Imperial" to "From Metric" for consistency.
Updated dev questions