ISSUE-231 Client: Fixed double score counting.
What was going on was that when submitting answers the results page wouldn't move forward until to show the scores until after the submission had been made. When that submission was complete there was a second pair of queries (one for Mastery and one for Survey) that would grab the "existing" scores. When writing this I didn't think it through. So that when the score board was shown it was grabbing the newly updated scores from the server. There isn't a race condition to worry about, the query only runs after the results have been submitted. Now instead I use the scoreChange prop and subtract it from the "current" score (I renamed it from "existing" score as it implies it is the pre-challenge score).