ISSUE-182 Client: Home now redirects to Welcome when user is logged out.
This effectively only works when a link is clicked. So if the visitor is visiting /home directly and isn't logged in, they'll still get a "you need to be logged in" notice. But when they hit the home button they'll go to /welcome.

Reduced the size of the mascot on the Welcome page by one notch.