ISSUE-250 Client: Added background art and favicons.
Included multiple flavors of the background pattern in 512x512 size in hard-coded opacities ranging from 1 to 7 and 100%.

Included favicons generated by realfavicongenerator.net.

Updates to the technology credits, removing the periods at the end of each quote.

Updated the UserDetailBasics page so that the list had a little more room to breath when in mobile mode.

Server: Updated language for Velocity subject.