ISSUE-262 Client: Created ScrollToTopOnMount component and updated Credits again.
I realized that when clicking on the credits page link it would be scrolled to the bottom. This makes it so that the scroll starts at the top every time the page is visited.

Idea taken from here: https://reacttraining.com/react-router/web/guides/scroll-restoration

Added ScrollToTopOnMount to the Subjects page as well.