ISSUE-320 Client: Added Welcome to the MenuFrame route.
This, in my opinion, will make the site more immediately interactive if a person wants to peek around, bored by the front page, before signing up for an account. It immediately communicates with a new visitor that this is a real website and there's a clear method to navigating it.

More random improvements to WelcomeText.