ISSUE-325 Server: Defined new docker-compose-aws-prod.yml file for the Server docker machine.
It doesn't use local host access for the database, instead relying on PRISMA_ENDPOINT settings in .env.aws.prod file for the DB machine.

Added a new env check for debug mode. Set the env var PRISMA_DEBUG to "false" (a string) to turn it off. Anything else (including unset) will leave it on.

Added a new start command that explicitly sets dotenv to use .env.aws.prod.

Defined new docker-compose-aws.yml file for the Prisma+DB docker machine.

The only difference is that it forwards port 80 traffic to 4466 where the Prisma API is listening. This allows one to skip the need for a port when making calls from the Server machine.