ISSUE-342 Client: Improved look on Tablets and extra wide browser windows.
Some text was cut off at the top of the page on tablets @768px.

On super wide screens the two mascot images would start to wander away from where it should be so I had to make an explicit fix.

The Carousel is still a little squished for @768px, but I'm just gonna let it slide for now...

Updating stats page a little bit.