ISSUE-341 Client: Saving my progress for day...
Very frustrating few hours of work as tutorials all over the Internet are out of date as Traefik has deprecated old methods and it's hard to find new methods.

Current issue is that the traefik.toml file isn't being read when bound through the volume. Not at all sure why this is happening.