ISSUE-341 Client: Working around docker-machine limitations with volumes.
Thing is that volumes copy files from the Host. And in the case of a docker-machine, the host isn't my laptop, but the EC2 instance. So there is no traefik.toml or acme.json file to mount. Unless I want to rsync with the instance, there is no easy method to get the files on there. So instead I have the Dockerfile copy the two files.

One problem I'm still trying to figure out is why the dashboard isn't working now.