ISSUE-341 Client: Switching strategy of how traefik.toml and acme.json files are handled.
Now uploading to the AWS hosts the Traefik and Acme files.

It would've been really nice to decide between SSL and non-SSL just by choosing a different docker-compose file, but I couldn't manage it. So I have to upload one file and touch a blank one to get it to work.