ISSUE-341 Server: Adding Traefik configurations for server.
Very similar to client. But there is one addition of a new dotenv file (.env.aws.prod.traefik) that is different in that it has the Prisma Endpoint to have an HTTPS address.

New package script "start-prod-traefik" to run when doing HTTPS.

Wrote a good script that backs-up the database.