ISSUE-350 Client: Figured out a way to solve /top edge case.
Before this little redirect trick, if someone scrolled to the bottom and hit the "Back to top" link they'd be directed to /all/top, which would auto-scroll to the top. But if they then immediately scrolled all the way back to the bottom the auto-scroll would no longer work.

So, as a work-around, in the case that it's still /all/top, the link will go instead to /all/up, which redirects immediately to all/top. It causes the page to jump up instead, which is a fine limitation considering the unusual case that it would be needed.