ISSUE-350 Client: Added some more colors to DocumentationMap component.
Added colors and an icon to To Metric and From Metric for the sake of visual recognition.

Narrowed the "Tips and Guides" Segment to be slightly thinner for the sake of appearance.

Updated DocumentationSubjectModal and the few other /docs links around the Client code.

Random Fix to FloatingCenterGrid and as a result also took the chance to update the NotFoundPage component.